Fixed resetSerialPortRcvBuf function which was resetting the flag allowing the serial port receive interrupt to process before resetting the associated variables, which would cause main code and interrupt to be conflicting with those variables.
Changed polarity of flag controlling AC OK LED so that a low turns it on
so that it is consistent with the other flags. The signal received in the
Main PIC from the power supply is actually high...that PIC inverts it
before transmitting.

Added define DEBUG_MODE	and renamed all usages to make it easier to find.
The define is commented out unless debugging in the simulator.

Signed-off-by: Mike Schoonover <mike@skoonie.com>